More tests on order

+++ src/main/java/com/example/demo/service/OrderService.java
@@ -25,7 +25,17 @@ public class OrderService {
         account.setPaymentHistory(currenPaymentHistory + 1);
     }
 
-    public void deleteOrder(long routeId) {
-        routeService.deleteOrderById(routeId);
+    public void deleteOrder(long accountId, long routeId) {
+        Account account = accountService.getAccountById(accountId).orElse(null); // Use orElse(null) to handle the case where the account may not exist.
+
+        if (account != null) {
+            routeService.deleteOrderById(routeId);
+
+            int currentPaymentHistory = account.getPaymentHistory();
+            account.setPaymentHistory(currentPaymentHistory - 1);
+
+        } else {
+            throw new InvalidInputException("Account not found for accountId: " + accountId);
+        }
     }
 }
